imm: ccbCreate - Fix expression for appending default values to ccb-create-callback [#1083]
The incorrect expression resulted in the omission of appending to the
ccb-create-oi-callback, the default values for attributes that (a) where not
explicitly assigned in the create request; and (b) did have a default value
defined.

The default was assigned in the imm-ram-db, but did not reach the OI
or appliers, including the PBE-OI and the special appliers (imcn).

The PBE (persistence) was actually no problem since a restart from the
PBE would again not assign these values resulting in the defaults being
set again in imm-ram.

If this bug caused any problem it could have been the omission to include
the attribute-value(default) in the callback to special applier/imcn.