IMM: Validating longDnsAllowed=0 checks normal RDNs are at most 64 bytes [#1009]
Updating the config attribute 'longDnsAllowed' in the object
'opensafImm=opensafImm,safApp=safImmService' to the value 0 has the
meaning of not allowing long DNs. But it also has the meaning of going
back to not allowing regular RDN attribute values longer than 64 bytes.

RDN attributes declared on the type SaNameT still allow their RDN values
to be longer than 64 bytes but not longer than 256 bytes (when longDnsAllowed
is zero). Such RDN attribute values are actually DNs and the objects are so
called association objects.

This patch fixes so that the validation for longDnsAllowed, when transitioning
from non-zero (long DNs allowed) to zero (long DNs not allowed) also covers
the restriction on regular RDN values.