immtools: Support for dumping instances of selected classes [#927]
Support both sqlite and xml.
Invalid class names will be ignored.
Epoch is not increased in case of dumping selected classes.
To make it possible to import the dumped xml file with immcfg, the objects will be sorted (parents come before children) and the 3 special attributes (SaImmAttrClassName, SaImmAttrAdminOwnerName, SaImmAttrImplementerName) will not be dumped.