plm: fix memory corruption in plms [#1244] Symptoms of the problem include PLMS crashing, and the node resetting because of the crash.
Valgrind shows memory corruption in the plms code. A double free, and some
uninitialized reads.

Solution for the double free is...don't double free! For the uninitialized
read the malloc'd string did not take into account the NULL terminator, and
functions later on (strchr) depend on a NULL terminator being there.