smf: implementer names reused [#1126]
Changed from using DNs as campaign and procedure OI names, to use fixed OI name for campaign ("safSmfCampaign"), and
a numbered name ("safSmfProc"<index>) for procedure OIs.
Extended the special handling where upgrade happens from and old version of SMF in which the campaign handles used also for procedures.
With this change it is possible to upgrade from that old version to this version.
The procedure OI names are now reused, since each campaign-start resets the index for procedure OI names.